+++ examples/with-ai-sdk-v5/app/page.tsx
@@ -3,7 +3,7 @@
 import { Thread } from "@/components/assistant-ui/thread";
 import { useChat } from "@ai-sdk/react";
 import { AssistantRuntimeProvider } from "@assistant-ui/react";
-import { useAISDKRuntime } from "@assistant-ui/react-ai-sdk-v5";
+import { useAISDKRuntime } from "@assistant-ui/react-ai-sdk-v4";
 
 export default function Home() {
   const chat = useChat();
