+++ packages/react-ai-sdk-v4/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@assistant-ui/react-ai-sdk",
-  "version": "0.11.0-alpha.0",
+  "name": "@assistant-ui/react-ai-sdk-v4",
+  "version": "0.10.16",
   "license": "MIT",
   "type": "module",
   "exports": {
@@ -22,16 +22,16 @@
     "lint": "eslint ."
   },
   "dependencies": {
-    "@ai-sdk/provider": "^2.0.0",
-    "@ai-sdk/react": "^2.0.0",
-    "ai": "^5.0.0",
+    "@ai-sdk/provider": "^1.1.3",
+    "@ai-sdk/react": "*",
+    "@ai-sdk/ui-utils": "*",
     "@radix-ui/react-use-callback-ref": "^1.1.1",
     "@types/json-schema": "^7.0.15",
     "assistant-stream": "^0.2.21",
     "json-schema": "^0.4.0",
-    "zod": "^4.0.14",
+    "zod": "^3.25.67",
     "zod-to-json-schema": "^3.24.6",
-    "zustand": "^5.0.7"
+    "zustand": "^5.0.6"
   },
   "peerDependencies": {
     "@assistant-ui/react": "^0.10.33",
@@ -46,11 +46,11 @@
   "devDependencies": {
     "@assistant-ui/react": "workspace:*",
     "@assistant-ui/x-buildutils": "workspace:*",
-    "@types/node": "^24.1.0",
-    "@types/react": "^19.1.9",
+    "@types/node": "^24.0.4",
+    "@types/react": "^19.1.8",
     "eslint": "^9",
-    "eslint-config-next": "15.4.5",
-    "react": "^19.1.1",
+    "eslint-config-next": "15.3.4",
+    "react": "^19.1.0",
     "tsx": "^4.20.3"
   },
   "publishConfig": {
@@ -59,7 +59,7 @@
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/assistant-ui/assistant-ui/tree/main/packages/react-ai-sdk-v5"
+    "url": "https://github.com/assistant-ui/assistant-ui/tree/main/packages/react-ai-sdk"
   },
   "bugs": {
     "url": "https://github.com/assistant-ui/assistant-ui/issues"
