+++ packages/react-ai-sdk-v4/scripts/build.mts
