+++ packages/react-ai-sdk-v4/src/dangerous-in-browser/DangerousInBrowserAdapter.ts
