+++ packages/react-ai-sdk-v4/src/dangerous-in-browser/createEdgeRuntimeAPI.ts
