+++ packages/react-ai-sdk-v4/src/rsc/VercelRSCMessage.tsx
