+++ packages/react-ai-sdk-v4/src/rsc/utils/RSCThreadExtras.tsx
