+++ packages/react-ai-sdk-v4/src/ui/getVercelAIMessages.tsx
@@ -2,8 +2,8 @@ import {
   getExternalStoreMessages,
   type ThreadMessage,
 } from "@assistant-ui/react";
-import type { UIMessage } from "ai";
+import type { Message } from "@ai-sdk/ui-utils";
 
 export const getVercelAIMessages = (message: ThreadMessage) => {
-  return getExternalStoreMessages(message) as UIMessage[];
+  return getExternalStoreMessages(message) as Message[];
 };
