+++ packages/react-ai-sdk-v4/src/ui/utils/useInputSync.tsx
