+++ packages/react-ai-sdk-v4/src/useChatRuntime.ts
