--- packages/react-ai-sdk-v5/src/index.ts
@@ -1,2 +1,0 @@
-export * from "./ui";
-export { frontendTools } from "./frontendTools";
