--- packages/react-ai-sdk-v5/src/ui/index.ts
@@ -1,1 +1,0 @@
-export { useAISDKRuntime } from "./use-chat/useAISDKRuntime";
