+++ packages/react-ai-sdk/src/frontendTools.ts
@@ -1,4 +1,4 @@
-import { jsonSchema } from "@ai-sdk/ui-utils";
+import { jsonSchema } from "ai";
 import type { JSONSchema7 } from "json-schema";
 
 export const frontendTools = (
