+++ packages/react-ai-sdk/src/ui/getVercelAIMessages.tsx
@@ -2,8 +2,8 @@ import {
   getExternalStoreMessages,
   type ThreadMessage,
 } from "@assistant-ui/react";
-import type { Message } from "@ai-sdk/ui-utils";
+import type { UIMessage } from "ai";
 
 export const getVercelAIMessages = (message: ThreadMessage) => {
-  return getExternalStoreMessages(message) as Message[];
+  return getExternalStoreMessages(message) as UIMessage[];
 };
