+++ packages/react-ai-sdk/src/ui/index.ts
@@ -1,4 +1,1 @@
-export { useVercelUseChatRuntime } from "./use-chat/useVercelUseChatRuntime";
-export { useVercelUseAssistantRuntime } from "./use-assistant/useVercelUseAssistantRuntime";
-export { getVercelAIMessages } from "./getVercelAIMessages";
-export { AISDKMessageConverter } from "./utils/convertMessage";
+export { useAISDKRuntime } from "./use-chat/useAISDKRuntime";
